requirements.txt
Django==4.0.4
djangorestframework==3.13.1
pandas==1.4.2
pyyaml==6.0
django-rest-swagger==2.2.0